switch numbers from short to float

+++ src/main/java/icl/ast/AstNum.java
@@ -1,7 +1,7 @@
 package icl.ast;
 
 public class AstNum extends AstNode {
-	public final short value;
+	public final float value;
 
 	public AstNum(short value) {
 		this.value = value;
